vm: fix some size_t/uint64_t issues (28)

+++ src/cshared/buffer.cpp
@@ -46,7 +46,7 @@ void Buffer::resize(uint64_t new_length)
     byte *new_data = (byte *)malloc((size_t)new_length);
     if (data)
     {
-        memcpy(new_data, this->data, min(this->length, new_length));
+        memcpy(new_data, this->data, (size_t)min(this->length, new_length));
     }
     free(this->data);
     this->data   = new_data;
